+++ apps/starter/modules/common/components/index.ts
@@ -1,0 +1,4 @@
+export { default as Layout } from "./Layout";
+export { default as Navbar } from "./Navbar";
+export { default as TerraWallet } from "./TerraWallet";
+export { default as WalletModal } from "./WalletModal";