+++ apps/starter/modules/common/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from "./useContracts";
+export * from "./useHive";