+++ packages/eslint-preset-arthuryeti/index.js
@@ -2,10 +2,12 @@ module.exports = {
   extends: ["next", "prettier"],
   settings: {
     next: {
-      rootDir: ["./apps/*/", "./packages/*/"],
+      rootDir: ["apps/*/", "packages/*/"],
     },
   },
   rules: {
     "react-hooks/exhaustive-deps": "off",
+    "@next/next/no-html-link-for-pages": "off",
+    "react/jsx-key": "off",
   },
 };