+++ apps/starter/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @arthuryeti/starter
 
+## 2.2.0
+
+### Minor Changes
+
+- Improving starter kit
+
+### Patch Changes
+
+- Updated dependencies
+  - @arthuryeti/terra@2.2.0
+
 ## 2.0.0
 
 ### Major Changes